+++ cmd/run/list.go
@@ -3,9 +3,11 @@ package cmd
 import (
 	"database/sql"
 	"errors"
+	"fmt"
 	"log/slog"
 
-	"github.com/rodaine/table"
+	"github.com/jedib0t/go-pretty/v6/table"
+	"github.com/jedib0t/go-pretty/v6/text"
 	"github.com/samcarswell/trochilus/config"
 	"github.com/samcarswell/trochilus/core"
 	"github.com/samcarswell/trochilus/opts"
@@ -13,10 +15,15 @@ import (
 )
 
 var nameOpt = "name"
+var statusField = "Status"
+var format string
 
 var listCmd = &cobra.Command{
 	Use:   "list",
 	Short: "Lists runs",
+	PreRunE: func(cmd *cobra.Command, args []string) error {
+		return opts.FormatTableOptValidate(cmd, format)
+	},
 	Run: func(cmd *cobra.Command, args []string) {
 		logger := slog.Default()
 		conf := config.GetConfig()
@@ -38,33 +45,95 @@ var listCmd = &cobra.Command{
 		if err != nil {
 			core.LogErrorAndExit(slog.Default(), err)
 		}
-		tbl := table.New(
+		var rows = []core.RunShow{}
+
+		for _, runRow := range runRows {
+			data := core.RunShow{
+				ID:            runRow.Run.ID,
+				JobName:       runRow.Job.Name,
+				StartTime:     core.FormatTime(runRow.Run.StartTime, conf.LocalTime),
+				EndTime:       core.FormatTime(runRow.Run.EndTime.Time, conf.LocalTime),
+				LogFile:       runRow.Run.LogFile,
+				SystemLogFile: runRow.Run.ExecLogFile,
+				Status:        runRow.Run.Status,
+				Pid:           core.FormatPid(runRow.Run.Pid),
+				Duration:      core.FormatDuration(runRow.Run.StartTime, runRow.Run.EndTime.Time),
+			}
+			rows = append(rows, data)
+		}
+
+		t := core.NewTable(rows, rowConv(conf), []string{
 			"ID",
 			"Job Name",
 			"Start Time",
 			"End Time",
 			"Log File",
 			"Exec Log File",
-			"Status",
-		)
-		for _, run := range runRows {
-			tbl.AddRow(
-				run.Run.ID,
-				run.Job.Name,
-				core.FormatTime(run.Run.StartTime, conf.LocalTime),
-				core.FormatTime(run.Run.EndTime.Time, conf.LocalTime),
-				run.Run.LogFile,
-				run.Run.ExecLogFile,
-				core.FormatStatus(core.RunStatus(run.Run.Status)),
-			)
-		}
-		tbl.Print()
+			statusField,
+		})
+		statusTransformer := text.Transformer(func(val any) string {
+			if status, ok := val.(string); ok {
+				color := text.FgWhite
+				switch status {
+				case core.FormatStatus(core.RunStatusSucceeded, conf.Display.Emoji):
+					if conf.Display.Color.Status.Succeeded {
+						color = text.FgGreen
+					}
+				case core.FormatStatus(core.RunStatusFailed, conf.Display.Emoji):
+					if conf.Display.Color.Status.Failed {
+						color = text.FgHiRed
+					}
+				case core.FormatStatus(core.RunStatusRunning, conf.Display.Emoji):
+					if conf.Display.Color.Status.Running {
+						color = text.FgCyan
+					}
+				case core.FormatStatus(core.RunStatusSkipped, conf.Display.Emoji):
+					if conf.Display.Color.Status.Skipped {
+						color = text.FgYellow
+					}
+				case core.FormatStatus(core.RunStatusTerminated, conf.Display.Emoji):
+					if conf.Display.Color.Status.Terminated {
+						color = text.FgHiMagenta
+					}
+				}
+				return color.Sprintf("%s", status)
+			}
+			return fmt.Sprint(val)
+		})
 
+		t.Table.SetColumnConfigs([]table.ColumnConfig{
+			{
+				Name:        statusField,
+				Transformer: statusTransformer,
+			},
+		})
+		t.Print(core.OutputFormat(format))
 	},
+}
+
+func rowConv(conf config.Config) func(core.RunShow, core.OutputFormat) table.Row {
+	return func(row core.RunShow, format core.OutputFormat) table.Row {
+		var status string
+		if format == core.FormatPretty {
+			status = core.FormatStatus(core.RunStatus(row.Status), conf.Display.Emoji)
+		} else {
+			status = row.Status
+		}
+		return table.Row{
+			row.ID,
+			row.JobName,
+			row.StartTime,
+			row.EndTime,
+			row.LogFile,
+			row.SystemLogFile,
+			status,
+		}
+	}
 }
 
 func init() {
 	RunCmd.AddCommand(listCmd)
 
 	listCmd.Flags().String(nameOpt, "", "Name of job to filter on")
+	opts.FormatTableOpt(listCmd, &format)
 }