+++ opts/opts.go
@@ -2,6 +2,7 @@ package opts
 
 import (
 	"errors"
+	"fmt"
 	"log/slog"
 
 	"github.com/samcarswell/trochilus/core"
@@ -31,4 +32,30 @@ func GetInt64OrExit(cmd *cobra.Command, name string) int64 {
 		core.LogErrorAndExit(slog.Default(), errors.New("option "+name+" must have a value"))
 	}
 	return optVal
+}
+
+func FormatTableOpt(cmd *cobra.Command, dest *string) {
+	cmd.Flags().StringVarP(dest, "format", "f", string(core.FormatPretty), "Format output (pretty|json|csv|tsv)")
+}
+
+func FormatObjectOpt(cmd *cobra.Command, dest *string) {
+	cmd.Flags().StringVarP(dest, "format", "f", string(core.FormatJson), "Format output (json)")
+}
+
+func FormatTableOptValidate(cmd *cobra.Command, format string) error {
+	switch core.OutputFormat(format) {
+	case core.FormatPretty, core.FormatCsv, core.FormatTsv, core.FormatJson:
+	default:
+		return fmt.Errorf("invalid format: %s", format)
+	}
+	return nil
+}
+
+func FormatObjectOptValidate(cmd *cobra.Command, format string) error {
+	switch core.OutputFormat(format) {
+	case core.FormatJson:
+	default:
+		return fmt.Errorf("invalid format: %s", format)
+	}
+	return nil
 }